♻️ Split core test case code into a `common` module
