+++ integration_tests/tests/common/mod.rs
@@ -1,0 +1,44 @@
+use std::path::{Path, PathBuf};
+
+use bintest::{BinTestBuilder, Command};
+use once_cell::sync::Lazy;
+
+#[derive(Clone, Copy)]
+pub enum TestBinary {
+    Compiler,
+    Interpreter,
+}
+
+pub struct TestCase {
+    pub command: Command,
+    pub input: PathBuf,
+    pub expected: PathBuf,
+}
+
+static TEST_CASE_DIR: Lazy<PathBuf> = Lazy::new(|| Path::new(std::env!("CARGO_MANIFEST_DIR")).join("test_cases"));
+
+pub fn get_test_case(binary: TestBinary, name: &str) -> TestCase {
+    let test_case_path = TEST_CASE_DIR.join(name);
+
+    TestCase {
+        command: get_test_binary(binary),
+        input: test_case_path.join("input.cat"),
+        expected: test_case_path.join("output.txt"),
+    }
+}
+
+fn test_binary_name(binary: TestBinary) -> &'static str {
+    match binary {
+        TestBinary::Compiler => "catastrophicc",
+        TestBinary::Interpreter => "catastrophici",
+    }
+}
+
+fn get_test_binary(binary: TestBinary) -> Command {
+    let bintest = BinTestBuilder::new()
+        .build_workspace(true)
+        .build_executable(test_binary_name(binary))
+        .build();
+
+    bintest.command(test_binary_name(binary))
+}
